backend/roe/requirements.txt
Django
openai
pydub
python-dotenv
numpy
faiss-cpu